Bain_ancova_cm {Bain}	R Documentation

Bayes factors of informative hypotheses for ANCOVA models (user-friendly hypotheses input)

Description

This function computes approximated adjusted fractional Bayes factors for ancova test with user-friendly hypotheses input.

Usage


Bain_ancova_cm(X, dep_var, covariates, group, hyp)

Arguments

X
A data frame that contains dependent variable, covariates, and group factor.

dep_var
A character that indicates the name of the dependent variable in X. Only one dependent variable can be specified.

covariates
A character vector that indicates names of covariates in X.

group
A character that indicates the name of the group factor in X. Only one group factor can be specified.

hyp
A character string that specifies hypotheses under evaulation. Variable names in the hypotheses have to be specified as 'group.n' with n = 1,2,3... denoting group numbers (e.g., group.1, group.2 and group.3), or 'group.category' (e.g., Sex.M and Sex.F for group factor 'Sex' with two categories 'M' and 'F'). Group/categpory names can contain: letters, numbers, and _. Variable names in one constraint have to be linked by +, -, =, >, <, e.g., group.1=group.2>group.3, or Sex.M<Sex.F. Different constraints in one hypothesis can be separated by "&", e.g., hyp = "Sex.F>0&Sex.M<0". Competing hypotheses can be separated by ';', e.g., hyp = "Sex.M=Sex.F; Sex.M>Sex.F". See examples and create_matrices for details.

Value

fit
Fit for each hypothesis under comparison

complexity
Complexity for each hypothesis under comparison

BF
Bayes factor for each hypothesis compared to its complement

PMPa
Posterior probability for each hypothesis excluding unconstrained hypothesis

PMPb
Posterior probability for each hypothesis including unconstrained hypothesis

Author(s)

Xin Gu, Herbert Hoijtink, Joris Mulder

Examples


  data(ChickWeight)
  #Hypotheses
  hyp<-"Diet.1=Diet.2=Diet.3=Diet.4;
        Diet.1<Diet.2<Diet.3<Diet.4;
        Diet.1<Diet.2<Diet.4<Diet.3"

  res<-Bain_ancova_cm(ChickWeight,"weight","Time","Diet",hyp) #run
  #Results are printed.
  #Results for fit, complexity, Bayes factor, and PMPs are saved in "res":

  plot(res)
  #Results for PMPs are plotted.


[Package Bain version 0.1.3 Index]