Bain_anova_cm {Bain}	R Documentation

Bayes factors of informative hypotheses for ANOVA models (user-friendly hypotheses input)

Description

This function computes approximated adjusted fractional Bayes factors for anova test with user-friendly hypotheses input.

Usage


Bain_anova_cm(X, dep_var = NULL, group = NULL, hyp)

Arguments

X
A data frame that contains dependent variable and group factor.

dep_var
A character that indicates the name of the dependent variable in X.

group
A character that indicates the name of the group factor in X.

hyp
A character string that specifies hypotheses under evaulation. Variable names in the hypotheses have to be specified as 'group.n' with n = 1,2,3... denoting group numbers (e.g., group.1, group.2 and group.3), or 'group.category' (e.g., Sex.M and Sex.F for group factor 'Sex' with two categories 'M' and 'F'). Group/categpory names can contain: letters, numbers, and _. Variable names in one constraint have to be linked by +, -, =, >, <, e.g., group.1=group.2>group.3, or Sex.M<Sex.F. Different constraints in one hypothesis can be separated by "&", e.g., hyp = "Sex.F>0&Sex.M<0". Competing hypotheses can be separated by ';', e.g., hyp = "Sex.M=Sex.F; Sex.M>Sex.F". See examples and create_matrices for details.

Value

fit
Fit for each hypothesis under comparison

complexity
Complexity for each hypothesis under comparison

BF
Bayes factor for each hypothesis compared to its complement

PMPa
Posterior probability for each hypothesis excluding unconstrained hypothesis

PMPb
Posterior probability for each hypothesis including unconstrained hypothesis

Author(s)

Xin Gu, Herbert Hoijtink, Joris Mulder

Examples


  #Example 1
  #Hypotheses
  #H1: group.1=group.2=group.3   #group means
  #H2: group.1>group.2>group.3
  #H3: group.1<group.2<group.3

  #Input
  X<-data.frame(c(rnorm(20,0,1),rnorm(40,.5,2),rnorm(60,1,5)),
                c(rep(1,20),rep(2,40),rep(3,60)))
  names(X)<-c("dep_var","group")
  head(X)

  hyp<-"group.1=group.2=group.3;
        group.1>group.2>group.3;
        group.1<group.2<group.3"

  res<-Bain_anova_cm(X,"dep_var","group",hyp) #run
  #Results are printed.
  #Results for fit, complexity, Bayes factor, and PMPs are saved in "res":

  plot(res)
  #Results for PMPs are plotted.

  #Example 2
  #Hypotheses

  #Input
  df<-data.frame(c(rbinom(20,10,0.5),rbinom(40,10,0.6),rbinom(30,10,0.6),rbinom(30,10,0.7)),
                c(rep("A",20),rep("B",40),rep("AB",30),rep("O",30)))
  names(df)<-c("ambition","bloodtype")
  head(df)

  hyp<-"bloodtype.O = bloodtype.AB = bloodtype.A = bloodtype.B;
        bloodtype.O > bloodtype.AB > bloodtype.A = bloodtype.B"

  res<-Bain_anova_cm(df,"ambition","bloodtype",hyp) #run
  #Results are printed.
  #Results for fit, complexity, Bayes factor, and PMPs are saved in "res":

  plot(res)
  #Results for PMPs are plotted.


[Package Bain version 0.1.3 Index]