Bain_regression_cm {Bain}	R Documentation

Bayes factors of informative hypotheses for regression models

Description

This function computes approximated adjusted fractional Bayes factors for hypothesis tests in regression models.

Usage


  Bain_regression_cm(formula, data, hyp, standardize = FALSE)

Arguments

formula
A formula that specifies the regression model. See formula or lm for details.

data
A data frame that contains the variables in the regression model.

hyp
A character string that specifies hypotheses under evaulation. Variable names should match the names of regression predictors. For example, if formula = Y ~ X1 + X2 + X3, variable names should consist of "X1", "X2" and "X3", but cannot contain e.g., "X4". Variable names in one constraint have to be linked by +, -, =, >, <. Different constraints in one hypothesis can be separated by "&". Competing hypotheses can be separated by ';'. For example, hyp = "X1>0&X2>X3;X1=X2=X3=0". See examples and create_matrices for details.

standardize
logical. If FALSE (default), then the estimates and covariance matrix of coefficients will not be standardized. If TRUE, then the estimates and covariance matrix of coefficients will be standardsized. When testing whether coefficients are zero, positive, or negative, we suggest standardize = FALSE. When comparing coefficients, we suggest standardize = TRUE. The standardization uses function "seBeta" in package "fungible", see reference for the theory.

Value

fit
Fit for each hypothesis under comparison

complexity
Complexity for each hypothesis under comparison

BF
Bayes factor for each hypothesis compared to its complement

PMPa
Posterior probability for each hypothesis excluding unconstrained hypothesis

PMPb
Posterior probability for each hypothesis including unconstrained hypothesis

Author(s)

Xin Gu, Herbert Hoijtink, Joris Mulder

References

Jones, J.A. and Waller, N.G. (2015). The normal theory and asymptotic distribution-free (ADF) covariance matrix of standardized regression coefficients: theorical extensions and finite sample behavior. Psychometrika, 80, 365-378.

Examples


  ##Example 1
  #Input
  data(cars)

  hyp = "speed = 0"
  res<-Bain_regression_cm(dist ~ speed, cars, hyp) #run
  #Results are printed.
  #Results for fit, complexity, Bayes factor, and PMPs are saved in "res":

  plot(res)
  #Results for PMPs are plotted.


  ##Example 2
  #Input
  data(LifeCycleSavings)

  hyp<-"pop15 = dpi = ddpi = 0;
        pop15 < 0 & dpi = 0 & ddpi > 0;
        pop15 < dpi < ddpi"

  #run
  res<-Bain_regression_cm(sr ~ pop15 + pop75 + dpi + ddpi, LifeCycleSavings,
                          hyp, standardize = TRUE)
  ## assume "pop75" is not in the hypothesis.
  #Results are printed.
  #Results for fit, complexity, Bayes factor, and PMPs are saved in "res":

  plot(res)
  #Results for PMPs are plotted.


[Package Bain version 0.1.3 Index]