Sensitivity analysis for bain

Conducts a sensitivity analysis for bain.

bain_sensitivity(x, hypothesis, fractions = 1, ...)

Arguments

x
An R object containing the outcome of a statistical analysis. Currently, the following objects can be processed: lm(), t_test(), lavaan objects created with the sem(), cfa(), and growth() functions, and named vector objects. See the vignette for elaborations.

hypothesis
A character string containing the informative hypotheses to evaluate. See the vignette for elaborations.

fractions
A number representing the fraction of information in the data used to construct the prior distribution. The default value 1 denotes the minimal fraction, 2 denotes twice the minimal fraction, etc. See the vignette for elaborations.

...
Additional arguments passed to bain.

Value

A data.frame of class "bain_sensitivity".

Details

The Bayes factor for equality constraints is sensitive to a scaling factor applied to the prior distribution. The argument fraction adjusts this scaling factor. The function bain_sensitivity is a wrapper for bain, which accepts a vector for the fractions argument, and returns a list of bain results objects. A table with a sensitivity analysis for specific statistics can be obtained using the summary() function, which accepts the argument summary(which_stat = ...). The available statistics are elements of the $fit table (Fit_eq, Com_eq, Fit_in, Com_in, Fit, Com, BF, PMPa, and PMPb), and elements of the BFmatrix, which can be accessed by matrix notation, e.g.: summary(bain_sens, which_stat = "BFmatrix[1,2]").

Examples

sesamesim$site <- as.factor(sesamesim$site)
res <- lm(sesamesim$postnumb~sesamesim$site-1)
set.seed(4583)
bain_sens <- bain_sensitivity(res, "site1=site2;
                                    site2>site5",
                                    fractions = c(1,2,3))
summary(bain_sens, which_stat = "BF.c")

#>   Fraction           H1       H2
#> 1        1 0.0005978540 115.9799
#> 2        2 0.0004227466 115.9799
#> 3        3 0.0003451712 115.9799

summary(bain_sens, which_stat = "BFmatrix[1,2]")

#>   Fraction        H1.H2
#> 1        1 0.0003015044
#> 2        2 0.0002131958
#> 3        3 0.0001740736