plot.Bain {Bain}	R Documentation

Pie chart for Bain result.

Description

This function plots pie charts for posterior model probabilities of candidate hypotheses.

Usage


## S3 method for class 'Bain'
plot(x,...)

Arguments

x
Result from Bain, Bain_ttest, Bain_ttestData, Bain_anova, Bain_anova_core, Bain_ancova, Bain_ancova_core or Bain_regression function.

...
Optional arguments passed to the plot function

Value

Plot a pie chart of posterior model probabilities if only one informative hypothesis is evaluated. Plot two pie charts of posterior model probabilities if more than one informative hypotheses are evaluated: one for posterior model probabilities excluding the unconstrained hypothesis; one for posterior model probabilities including the unconstrained hypothesis.

See Also

Bain,plot

[Package Bain version 0.1.3 Index]