Simulated Sesame Street Data

This is a simulated counterpart of part of the Sesame Street data presented by Stevens (1996, Appendix A) concerning the effect of the first year of the Sesame street series on the knowledge of 240 children in the age range 34 to 69 months. We will use the following variables: sex; site of child's origin; setting in which Sesame Street is watched; age; whether or not the child is encouraged to watch; Peabody metal age score; score on numbers test before, after and in a follow up measurement; and scores on knowledge of body parts, letters, forms, numbers, relations, and classifications, both before and after watching Sesame Street for a year.

data(sesamesim)

Format

A data frame with 240 rows and 21 variables.

Details

sex	integer	Sex of the child; 1 = boy, 2 = girl

site	integer	Site of the child's origin; 1 = disadvantaged inner city, 2 = advantaged suburban , 3 = advantaged rural, 4 = disadvantaged rural, 5 = disadvantaged Spanish speaking

setting	integer	Setting in which the child watches Sesame Street; 1 = at home, 2 = at school

age	integer	Age of the child in months

viewenc	integer	Whether or not the child is encouraged to watch Sesame Street; 0 = no, 1 = yes

peabody	integer	Peabody mental age score of the child; the higher the score the higher the mental age

prenumb	integer	score on a numbers test before watching Sesame Street for a year

postnumb	integer	score on a numbers test after watching Sesame Street for a year

funumb	integer	follow up numbers test score measured one year after postnumb

Bb	integer	Knowledge of body parts before

Bl	integer	Knowledge of letters before

Bf	integer	Knowledge of forms before

Bn	integer	Knowledge of numbers before

Br	integer	Knowledge of relations before

Bc	integer	Knowledge of classifications before

Ab	integer	Knowledge of body parts after

Al	integer	Knowledge of letters after

Af	integer	Knowledge of forms after

An	integer	Knowledge of numbers after

Ar	integer	Knowledge of relations after

Ac	integer	Knowledge of classifications after

References

Stevens, J. (1996). Applied Multivariate Statistics for the Social Sciences. Mahwah NJ: Lawrence Erlbaum.