Student's t-test

This function is a wrapper for the function t.test, which returns group-specific sample sizes and variances, in addition to the usual output of t.test.

t_test(x, ...)

Arguments

x
An object for which an S3 method of t.test exists (vector or formula).

...
arguments passed to t.test.

Value

A list with class "t_test" containing the following components:

statistic

the value of the t-statistic.

parameter

the degrees of freedom for the t-statistic.

p.value

the p-value for the test.

conf.int

a confidence interval for the mean appropriate to the specified alternative hypothesis.

estimate

the estimated mean or difference in means depending on whether it was a one-sample test or a two-sample test.

null.value

the specified hypothesized value of the mean or mean difference depending on whether it was a one-sample test or a two-sample test.

alternative

a character string describing the alternative hypothesis.

method

a character string indicating what type of t-test was performed.

data.name

a character string giving the name(s) of the data.

v

The variance or group-specific variances.

n

The sample size, or group-specific sample size.

Details

This wrapper allows users to enjoy the functionality of bain with the familiar interface of the stats-function t.test.

For more documentation, see t.test.

See also

Examples

tmp <- t_test(extra ~ group, data = sleep)
tmp$n

#> [1] 10 10

tmp$v

#> [1] 3.200556 4.009000

tmp2 <- t_test(extra ~ group, data = sleep)
tmp2$n

#> [1] 10 10

tmp2$v

#> [1] 3.200556 4.009000

tmp <- t_test(extra ~ group, data = sleep, paired = TRUE)
tmp$n

#> [1] 10

tmp$v

#> [1] 1.512889

tmp2 <- t_test(extra ~ group, data = sleep, paired = TRUE)
tmp2$n

#> [1] 10

tmp2$v

#> [1] 1.512889

t_test(sesamesim$postnumb)

#> 
#> 	One Sample t-test
#> 
#> data:  sesamesim$postnumb
#> t = 36.244, df = 239, p-value < 2.2e-16
#> alternative hypothesis: true mean is not equal to 0
#> 95 percent confidence interval:
#>  27.84932 31.05068
#> sample estimates:
#> mean of x 
#>     29.45 
#> 

tmp <- t_test(sesamesim$prenumb)
tmp$n

#> [1] 240

tmp$v

#> [1] 112.728

tmp2 <- t_test(sesamesim$prenumb)
tmp2$n

#> [1] 240

tmp2$v

#> [1] 112.728

tmp <- t_test(sesamesim$prenumb, sesamesim$postnumb)
tmp$n

#> [1] 240 240

tmp$v

#> [1] 112.7280 158.4577

tmp2 <- t_test(sesamesim$prenumb, sesamesim$postnumb)
tmp2$n

#> [1] 240 240

tmp2$v

#> [1] 112.7280 158.4577

tmp <- t_test(sesamesim$prenumb, sesamesim$postnumb, paired = TRUE)
tmp$n

#> [1] 240

tmp$v

#> [1] 90.61583

tmp2 <- t_test(sesamesim$prenumb, sesamesim$postnumb, paired = TRUE)
tmp2$n

#> [1] 240

tmp2$v

#> [1] 90.61583